Refactor command modules. separate app definition and runner. (for visibility)
